Docs: Fix linking to inherited members with reified generics.
This is a confusing situation where the parameter type has constructed generics that are not generic on the method itself - namely, `RegisterDecoratedProperty` is only generic on `TValue`, but the `DecoratedProperty<T, TValue>` picks up the `T` from the calling class.

The solution in place _should_ be able to reliably link to the proper "base definition" without breaking intra-class links or other generic member links that are actually supposed to have reified generics, e.g. something referring to a `List<int>`.