Change `DecoratorView` to expose its view as a child.
Making the inner view "transparent" was causing problems with focus and other aspects when certain operations, such as overlays, became associated with that inner view.

Initially this was detected on the `DynamicDropDownList`, as the overlay was created by the inner `DropDownList<T>`, and when the `ViewMenu` saw this overlay, it could not find the `DropDownList<T>` that created and thus could not trace a path for layout or focus.

May also have been an issue with other overlay views like keybinding; did not get around to testing them before applying this fix.