Fix caret size and orientation in `Expander`.
Original code was likely bugged due to an old, incorrect and eventually corrected implementation of `SimpleRotation` for `Image`.

Also, forcing the caret to stretch to height tended to cause it to resize itself oddly, because the caret is not square. It looks better to use a fixed size, i.e. just set to the sprite's content size, and scale the sprite itself if need be.