Relax aggressive layout breaking when `Lane` runs out of space.
This was originally done as an optimization, but by suppressing layout on the entire subtree, it can actually cause severe thrashing by preventing the dirty flag on a descendant from ever being reset, resulting in a new layout every frame and an essentially unresponsive UI. This was reproducible with a simple Expander example.

It might be possible to re-optimize this by skipping subtree layout if that part of the tree is not dirty, but without other design changes, this would cause inconsistencies in the visible child list/count, and is probably not worth the effort to try to get perfect performance since any _correct_ layout should not normally have lanes running out of space anyway.