Add converter factory for nullable types.
These conversions are mostly safe - nullable to non-nullable is somewhat risky but users should still be able to attempt it.

Implicit assignments to nullable types are allowed at compile time but the reflection code doesn't understand them, so we have to be explicit. This comes up infrequently, but it does come up, e.g. in the `KeybindType?` (nullable enum) on the `KeybindListEditor`.