Move button sprite maps into Framework mod.
This was always the intent - part of the reason for creating a framework mod in the first place - but simply was forgotten/not done.

Adds asset handlers to the Framework mod so that these can be referenced in StarML using standard asset bindings; in the case of the `ButtonSpriteMap`, with or without parameters that allow tweaking theme selection and scaling.