Add duck type conversions.
Should work with quite a wide variety of constructor and property combinations, if not all (difficult to test exhaustively).

Includes refactoring to extract `RootValueConverter` and expose default ctor (without any registrations) for tests. Also extracts a non-generic `IValueConverter` which is pretty essential for this particular converter, though it shouldn't be needed in many other places.

Annotates types that are expected to need duck type conversion (`Sprite`, `Edges`, `LayoutParameters`, etc.)

#21