Add "re-hover" flag and trigger for menus.
Makes most input events (_except_ mouse moves, which don't need it) set a flag that will force the hover path to be reevaluated after a 1-frame delay.

This ensures that tooltips and any other hover state are updated when the path might have changed, even if the pointer hasn't technically moved.