Integrate resolution scope into view nodes.
This plumbs it all the way into the `AttributeBindingFactory`, where it is not actually implemented yet but will need to be used when resolving literal attribute values.

The parallel hierarchies _should_ be able to detect the default scope and also handle `<include>` views, although tests at this point are only to verify that no existing functionality is broken.

#28