Integrate `IResolutionScope` into value sources.
This is the last stage of the plumbing and should pass it all the way down to the `IValueSourceFactory.GetValueSource` method for attributes (not supported for event args).

Adds some unit tests for scoping, which can't quite cover E2E due to not having a real mod graph, but verified the results with in-game testing and an updated example, which uses both local scope (from TestMod) and qualified scope (settings key in Framework mod itself).

#28