Add shadows to Label widget.
Includes a `ShadowLayers` enum to try to mimic what vanilla does without introducing the same confusing API. Updates the `Banner` to support a consistent set of properties, although the `SpriteText` used in banner text already has a built-in shadow.

Adds a new `ShadowColor` property to be able to control the base color of the shadow, separately (and in combination with) the alpha.

For backward compatibility, the alpha property keeps its default of 0 and the layers default to `All`; the latter is not consistent with the default in previous StardewUI iterations but it **is** consistent with vanilla defaults.

#37