Rework layout offsets into base `View`.
Moves the localized scroll-offset behavior used in the `ScrollableContainer`, as well as the demo `Carousel`, into a property held by the `View`, where the scrollable or offsettable subclass simply reports its layout offset for most operations and only applies it specifically during drawing.

This makes the implementation of some methods simpler, despite feeling less intuitive overall. However, and most crucially, it allows the `View` to detect when that offset has changed during a pointer move event, thereby correctly dispatching enter/leave/move events to descendants even when the physical cursor position on screen technically hasn't changed at all, i.e. when scrolling or animating the carousel.

In addition, reworks a bit of the rehover logic in the `ViewMenu` to deal with the haphazard vanilla flow where both `performHoverAction` and `update` are called at seemingly arbitrary times, so the final outcome depends on which event actually triggered rehover. To make sure we only cancel the "previous request", we can use the game tick on which it was originally requested.