Switch to optional, default-off laziness for descriptor members.
With members being initialized on parallel/background threads, the advantage very likely turns to eager loading on all but the lowest-end machines, because most threadpool threads in Stardew are never busy, and lazy initialization means waiting until bind time to initialize the member descriptors, which forces them to happen on the main thread.

In other words, eager initialization maximizes the amount of work done _off_ the main thread, which is normally what we want. However, it could have problems if a type is really enormous, so the option to be lazy is kept in for now, even though it's no longer used.

#44