Add source generator for precompiled descriptors.
This is especially useful for View Descriptors, since all the standard views are known ahead of time, and in theory the source generator could also be run on add-on views.

Currently does not support opting in arbitrary objects (e.g. marker attribute can't be added to a `ParsedItemData`, we don't own it) but that is planned for the future using some other mechanism to trigger generation.

#18 #44