Performance tweaks for binding and asset loading.
The big fish here is asset loading; it turns out that loading content explicitly by asset path will incur the full cost of a file load, texture decode, etc., even if we already have a copy of that asset in `Game1`. Therefore, redirecting those requests to the relevant `Game1` asset can result in a major speedup (more than 50 ms for an image like Cursors).

Other tweaks include switching the somewhat expensive `GroupBy` in the parallel descriptor loading to `Distinct`, which is still supported for parallel, and _removing_ parallelism from `ReflectionViewBinder` because it was turning out to be slower, probably due to the low number of actual bindings per view and overhead of ordered parallelism.

#44