Handle empty literal attributes.
The error here is actually in the parser, not the lexer, as the parser was not expecting a closing quote token directly after the opening quote.

Note: We still do not allow empty binding-style attributes, as an empty binding makes no sense. Literals can be empty since they are naturally string-typed.

#43