Don't try to load missing localized assets.
In the `AssetRegistry`, we receive a valid request for a potentially translated asset, e.g. `viewname.fr-FR.sml`. While this isn't the preferred way of doing translations (better to use translation bindings), it is in fact possible and supported to have localized views, so the solution here is simply to check that the file really exists before trying to load it.

We also get a not-so-valid version of this in the main `AssetRequested` handler for built-in UI sprite names, where the loader appears to be appending a language code to the actual asset name provided in the view; so in this case, we should simply use the non-localized name because the in-memory asset should already be localized.

Fixes #47