Filter `ByRef` types from reflected descriptors.
These parameters don't have the `IsIn` or `IsOut` flags, they look like normal parameters but have a different type. This is probably due to different by-ref behavior for value types vs. reference types.

This was a recent regression because of how the reflector now eagerly loads all properties on parallel/background threads instead of lazily on the main thread (#44).

Fixes #54