Add reflection warmups.
This has some, albeit not huge benefit of around 10-15 ms on the formerly 120+ ms first-time load (now closer to 100ms in optimal conditions).

There are two types of "warmup", one is to actually place generic method delegates directly into the cache for very common/well-known types and conversions, which should be faster during regular UI loops as well as the first time. The second is mainly to "prime" the `MakeGenericMethod` and `MakeGenericType` methods which seems to have some internal caching going on in .NET, and runs faster after the first time.

Individually the improvements are all very tiny but they do add up to about 1 less jank frame.

Finally, this adds configuration settings for both reflection warmup and descriptor parallelism, and sets their defaults to what has been producing the best results so far; with precompiled descriptors now supported, parallel seems to be better OFF as there is not enough work to overcome the overhead.

#18 #44