Invalidate descriptors on metadata update.
Registers a `MetadataUpdateHandler` to hook hot reloads (from the debugger), invalidate the descriptor cache for any changed types, and invalidate any other objects known to depend on descriptors - currently just `DocumentView`, where a descriptor update gets treated similarly to an asset reload.

Tested by adding a property to the example model with debugger running, pause/resume to trigger hot reload, and then adding an element referencing the new property to the StarML, thereby triggering StardewUI's asset-based hot reload, and ensuring that the view renders correctly instead of throwing a descriptor exception.

Note that changing the type does not update live objects, and this is a limitation of hot reload itself; that is, adding a new field/property with a default value will not actually assign that default value to existing instances of that type; they have the new member, but not its default. To "warn" users of this, the handler logs an info message each time explaining that UI may need to be closed and reopened.

Resolves #45