Propagate structural attributes to templates.
Regular property attributes shouldn't be copied because the correct way to consume them is with template parameters. However, it's very confusing if structural attributes like `*repeat` or `*context` behave this way.

The behavior of structural attributes (with exceptions as noted) should be that they are copied to the root elements of each template, whether replaced or added, and then dropped from any descendants. Structural attributes are supposed to apply to _that specific node_ so they don't really make sense as parameters, and the expect behavior of something like a `<foo *repeat=...>` is obviously that the expanded template repeats.

#39