Rework test mod as real demo.
F8 now brings up a nice little animated menu with navigation into all the different examples, which open as submenus.

Adds a Form example showing an AFS-like config UI. Especially for demonstrating `<template>` and the `IMenuController`. (The test form already existed, this just turns it into a real example).

Note: Carousel link isn't working yet since that is added by the Addon. Need to figure out a way to patch this in.

#39