Account for `MinWidth` in layout.
`LayoutParameters.GetLimits` was only constraining to the `MaxWidth` (and height) and ignoring `MinWidth` (and height), so as a result, the outer layout size (e.g. frame) was the true width after constraints, but the inner content size would use the preferred width even if it was too small.

Swapping the `Min` (with max dimension) to a `Clamp` (with both min and max dimensions) fixes that issue.