Fix snap-back from text entry menu.
Uses the same mechanism that is used to refocus after closing an overlay, but also requires a fix to the capturing in `TextEntry` to actually exit capture mode when the menu is closed, since `Game1.closeTextEntry` doesn't give any kind of notification, and the standard draw/update hooks won't reliably fire due to how Stardew's update loop is implemented, so we need to use global updates.

Also removes the `Game1.lastCursorMotionWasMouse` check from the `TextInput`, which was already removed from other parts of the `ViewMenu` but left in the `TextInput` accidentally. The result of this was that opening the text entry menu (with gamepad), then exiting it, then immediately pressing A again, would be treated as a regular mouse click and not open the menu a second time. This is now fixed.