Forward close button events to title submenus with custom close actions.
`TitleMenu` does not forward them properly because `readyToClose` always returns `false` when custom, and the title menu has a hard-coded branch for that case.

Moves the title-menu check for custom close actions to after the custom action check, so that custom actions can also override title submenu behavior, which is required for the GMCM hack, which is currently the primary use case.