Use character breaking in Labels when single words exceed line limit.
The situation can apply to all languages including English/Latin-based (as tests show) but is particularly problematic in CJK where entire blocks of text may be written without any spaces, instead using full-width punctuation.

"Algorithm", such as it is, will hyphenate for Latin languages but not CJK, and attempt to avoid breaking on punctuation characters.

Includes a new test fixture that loads the game assembly in a similar fashion to StardewXnbHack, which appears to actually be the most straightforward way to create objects like `SpriteFont` and `Texture2D`, other workarounds requiring Windows Forms or other OS-specific methods of creating a window handle.

Fixes #73