Rewrite text input to use one label with scroll container.
This makes it behave a lot more like a regular HTML textbox that scrolls with the caret/cursor. There is no more ellipsis, and any content can be accessed by clicking or using the arrow keys. As a side effect, this also fixes the "wiggling" of text when the cursor moves because there is now only one label, not two.

Scroll container provides only part of the functionality here since it is really meant to scroll over many views, not a single view; however it does do a lot of what we want, especially handling the layout and clipping, clamping scroll offsets, etc. The main things that need to be reimplemented are peeking and adjusting for click positions.

Fixes two other undocumented bugs discovered in the process:

- Scrollable containers weren't accommodating for their own margins or padding in clipping behavior. The clip rect needs to be adjusted for negative margins and also allow content to overflow into padding (but not margins).
- Virtual keyboard was a little broken with controllers because the `ViewMenu` kept trying to refocus, probably after some other focus fix that went into 0.4.x. Updated to ignore refocusing when a keyboard subscriber (capture target) is active.

Fixes #66