Track previous pointer state in fields.
The reason pointer-enter/leave events weren't consistently firing was due to the previous position being the same as the current position when the first pointer event was received (i.e. when menu was opened, or even when a view is newly created/made visible in an active menu).

The simplifying assumption was that position always corresponded to hierarchy, but this was too simple because (a) hierarchy can change and (b) there is no "previous" hierarchy on the first event.

Reasons for the "inferred previous state" don't seem to be documented in the original commit, but may have been trying to avoid a memory leak or just avoid extra state. If so, the use of `WeakViewChild` eliminates that risk.

This can be tested in the demo mod. After the change, pressing F8 to open the menu automatically highlights the selected gallery item, and clicking on the "Form Example" such that the cursor is within the bounds that the keybind editor will appear, causes the keybind editor to do its hover immediately on showing.

This *may* also help with the issue of `PointerLeave` events getting lost under specific and unclear circumstances, but there was never a clear test case for it.

Fixes #62