Docs: Downrank reference pages.
Note, boost is not a sort order, it's a multiplier, so this doesn't guarantee that *all* reference pages appear below *all* guide pages. There seems to be some diminishing returns on very low levels, but the results appear mostly acceptable; for example:

- Searching for "context" brings up the Binding Context and Overlay Context pages first, then Context-related classes, then (with a custom extra boost) the StarML page, then other less-relevant classes and guide pages that just happen to reference context in an incidental way, like in a code example.
- "binding" brings up StarML first, then Binding Events, Binding Context, Concepts page, and then ref docs for Binding-related classes. Templates, Keybind, and many other guide docs are next, and then after those, the more esoteric Framework binding classes and interfaces.
- "Label" brings up Standard Views first, then StarML ref, then a few guide pages that aren't the _best_ for relevance, then the `Label` ref doc, then a lot of other less relevant results.

There isn't going to be a single perfect ranking system for every query - have simply tried to pick an OK balance for reference pages, with the expectation of doing a bit of boosting and manual ranking on particularly important guide pages.