Enable floating elements in StarML with `*float`.
Includes conversions to `FloatingPosition` from strings and functions, the latter to support duck type conversions for e.g. repeating floats. All covered in tests.

Floating elements are technically part of `View`, not `IView`, but always have a consistent type, so we can define an interface instead of relying on descriptors/reflection/etc. Special views (e.g. decorator views) must implement it explicitly.

Implementing it is primarily a matter of passing `FloatingElements` through the various node types similar to how `Views` are passed, and making sure they are compared and propagated wherever views would be compared or propagated. The complexity is all in the `ViewNode`, so as to be able to handle different binding types on the `*float` attribute like we do for `*context`.

Fixes a bug in `IValueConverterFactory` default `GetConverter` implementation that would throw an exception (only expected for "required" converters) instead of returning `null` as per spec. Also make `ConvertedValueSource` one-way by default to avoid the overhead, error logging, etc. of failed backward conversions, when this is almost never necessary - none of the existing instances of `ConvertedValueSource` appear to be used in a situation where two-way conversion could be meaningful.

Closes #67