Add whole-grid alignment for grids.
This will align the entire block of items (like table alignment in HTML) to help accommodate for extra space before or after.

It's not possible to accomplish this with general widget layout (e.g. sticking inside a `Panel` with `HorizontalAlignment` of `Middle`) because the grid's own layout size is independent from the accumulated size of the columns (or rows, in vertical orientation). Thus we need one property on the grid itself to do this alignment, along the primary axis, but not two, because the secondary axis does not have this discrepancy.

#64