Add global transforms with transform propagation.
The original `Transform` is now specifically a local transform, and `PropagatedSpriteBatch` uses the new `GlobalTransform`, which combines a `Matrix` with the local transform in order to try to preserve XNA's optimizations for trivially-combinable transforms.

Since the global transform is applied lazily in order to avoid thrashing on `SpriteBatch` begins/ends, we now require `ISpriteBatch` to implement `IDisposable` and for the `ViewMenu` to dispose it in order to make sure the transform is actually reverted at the end of StardewUI's event loop. Otherwise, extremely strange things can happen when code independent from the transform propagation keeps trying to use a transformed batch.

Note that this does **not** work with scissor rectangles, which unfortunately seem to be fixed to the screen/viewport and ignore any active transform.

#59