Implement transform origins.
This is very simple on the matrix multiplication side for global transforms, and very complicated for simple local transforms passed to drawing commands because of the awkward way MonoGame (inherited from XNA) handles origins. Because of the complexity, the `origin` argument has been removed from the `ISpriteBatch` wrapper's method arguments to avoid an even bigger mess of conflicts and confusion.

`NineSlice` (the only former user of the `origin` parameter) has to be rewritten, but the rewritten implementation is now much cleaner.

#59