Update scissor rectangles to use both scale and transform.
`ScissorRectangle` is relative to the viewport, so there's no way to use that specific API for a rotated container, but it _can_ still work with a scaled container since the result is still a rectangular region, and doing so is much more efficient than using another `RenderTarget2D` when it is possible.

The fallback implementation still needs to be updated to handle rotated/skewed transforms using a render target but this will optimize a large number of possible transform+clip scenarios.

#59