Use render targets for rotated clip regions.
When we have a "non-rectangular" region - skewed, rotated, otherwise incompatible with `ScissorRectangle` - use a `RenderTarget2D` instead to _create_ a rectangular drawing surface and then draw the rendered texture later with the proper transform.

This is not as efficient as `ScissorRectangle`, but is probably the best we can do. To try to make this as fast and seamless as possible, this uses a pool of `RenderTarget2D` instances to minimize the creation of new textures. If the pool is associated with e.g. a menu (the long-lived owner of the shorter-term `PropagatedSpriteBatch`) then these targets can be continually recycled.

Fixes some ordering issues with the reverters; mainly, we need to make sure to revert (or set) the transform _before_ beginning a new sprite batch, since `BeginSpriteBatch` is where we "apply" the transform.

#59