Support custom/optimized opacity implementation per view.
Simple views such as `Image` and `Label` don't need an entire render target in order to display transparently - they can simply adjust their colors. This can make animations much faster on these view types if animated in isolation.

Similar logic can be applied to certain composite views like `CheckBox`; however, this is limited to views with no overlapping components. When there is overlap, e.g. with a `Banner` comprising the "scroll" background and text, the outcome of drawing 2 opaque overlapping components and making the entire layer transparent is *not* the same as drawing each with individual transparency. Due to the default alpha blending, in the latter case much more of the content underneath the transparent view will bleed into the final result.

#59