Behaviors: Add syntax to grammar.
Uses the `+` prefix. These attributes are not handled anywhere yet, only added to the grammar.

Also updates to allow the colon character `:` in names, to support static behavior arguments. This will technically allow them in tag names as well, but should be OK since it is not ambiguous with any other syntax.

#68