Rework/rewrite of behavior system.
The basic idea and specification is intact, but the inputs to a behavior are now expanded to support "view states" (property overrides), and the API has been changed to incorporate the now-obvious finding that the view type needs to be known at behavior creation time in order to figure out what (for example) a `hover:transform` attribute is actually supposed to operate on, in terms of the expected value type, which we need in order to create a generic behavior type, or else make everything non-generic and suffer even more when trying to update the view.

Because the view and state overrides are almost always concurrent requirements, this creates a record type and passes them around generally as a single entity, right up until the behavior subclass where they are implicitly separated only in the property getters.

#59