Add data types for transitions and interpolations.
We already had `Lerp<T>`, but it was extremely bare bones. This adds a registry for per-type lerp functions, and adds Easing functions with the same system, as well as the actual Transition type.

While the core library doesn't exactly "use" these, they don't depend on any framework features either, so they make more sense to put in Core.

#59