Fix margins/padding for item grid example.
May not matter that much in a demo, but it was especially distracting with the scale-up animations being clipped by the scrollable boundary. Padding on the border frame should be only just large enough for the content and border not to overlap, with any additional spacing being inside the scrollable.

#59