Fix scrollbar clipping.
For some historical reason, the `ScrollableView` was initializing its `Scrollbar` to have a fixed width of 32px... when the actual default images used in the scrollbar have a width of 48px. This was causing them to draw outside the view's bounds, causing a variety of problems including clipping when used with transforms/opacity, and not actually receiving clicks properly near the right edges.

Fixing this also appears to fix the issue where the arrow's hover state would not reset when moving the cursor out of bounds on the right side, or vice versa.

Secondary update is to the floating bounds calculation which should take into account margins (since they are incorporated into the drawing itself) but previously were not.