Add a visibility behavior.
Can be used to apply property changes and therefore transitions when a view is shown, or made visible.

Doesn't really work for hiding views because once the view is hidden, it stops drawing entirely, therefore any transitions won't be displayed. A future improvement might find a way to add a `Visibility` override state until the transition is done, although this would still only work for `Hidden` views and not views that are completely removed.

In this regard it is almost exactly like CSS transitions, which don't have a clean way to do "removal animations" either, although many of the more advanced web animation frameworks do.

Updates the main Gallery example to use transition behaviors and removes all the viewmodel hacks.

#59