Prevent full layout for common sprite animations.
In the specific case where the sprite size doesn't change, which is true for almost all sprite animations, the layout can't change as a result of the sprite changing. We can therefore skip the normal layout and dirty check, and wait until the post-update to recreate the nine-slice.

This improves steady-state performance somewhat but also makes layout debugging much less painful since there are not a bunch of "useless" layouts constantly happening and triggering breakpoints.