Filter actual `null` pending values in `DropDownList` setter.
The previous logic was broken for value types, including enums, as it would never allow the option to be set back to the default value for that type (e.g. the number 0, the first enum value, etc.).

We need an _explicit_ boxed check for `null` because `default` for a value type is non-null.

Also cleans up a few issues in the example form itself, including a lack of two-way bindings.

#82