Supress controller-triggered menu close when an overlay is open.
The controller-button logic in this case short-circuits the key-press logic which used to handle that state, so we need to handle it in the short-circuited version too.

Adds some conditions to `receiveKeyPress` to try to prevent duplication of overlay pops and other fall-through.

Removes the explicit `receiveGamePadButton` from the main input hook as it seems to be duplicating the button presses. Unsure if this was a recent change in Stardew or if it was always duplicating.

Have tested using AFS, as a `TitleMenu` sub-menu (via GMCM), as a standalone menu (via mod keybinding) and as a child menu of GMCM in-game, the behavior now appears to be consistent between all three.

#83