Convert `Frame` content to `ViewChild` before running focus search.
This was the main discrepancy from all other view implementations, and was causing both the internal lookup position to be wrong as well as the final outcome to have the wrong offset.

In addition, while probably not strictly necessary, the `DecoratorView` focus search was not returning itself in the path, which was inconsistent with the `GetChild*` functions and making debugging very confusing, so this has been fixed as well.

#63