Patch `SpriteText.drawString` to ignore viewport boundaries while drawing UI.
Whatever the original rationale was for this constraint in vanilla, it doesn't really make sense for anything StardewUI does, especially when global transforms are active and the coordinates provided to the underlying `SpriteBatch` are not necessarily viewport-relative.

Tries to avoid letting the patched version run when invoked outside StardewUI, in case it really does have some important purpose in vanilla. Since failure to patch isn't really a fatal error (would just cause some visual glitches), this is wrapped in a continuing exception handler.

N.B. The test case for this is the Carousel example, when scrolled to the second page or later. Without this patch, `SpriteText` tries to render the title of the previous frame "on-screen" which gets translated to "within the clipping rectangle" (instead of outside it) because the clipped view has negative translation.