Improve stability of view states and behaviors across rebinds.
First, there was a blatant bug in `ViewBehaviors` that simply didn't recreate OR retarget an already-existing behavior in `SetTarget`. However, even with this fixed, several problems have come up in "real-world" use. Observed effects include:

- `+show:opacity` and other "transition into view" scenarios not showing the view at all after a high-level context is changed
- `+transition:` behaviors trying to transition back to much older and no longer valid states
- Transitions related specifically to `show` properties rerunning even when the view hasn't been recreated.

Fixing this involves a moderate rework of the behavior implementation to allow reuse and "reinitialization" of behaviors which translates into explicit attach/detach instead of the previous `OnInitialize`/`OnDispose` (though the latter still exists). In addition, when creating a new `ViewState`, the state lists associated with the previous state are preserved, so that only the default values and not the current states change.

This seems to be enough for the scenarios being tested so far; more changes may come later.