Remove generic parameter from `ViewMenu`.
Not a single line of code in the core library or framework cares what the view type is, so it's unnecessary overhead that makes debugging needlessly painful (since hot reload does not work with generic types).

In the rare case that some subclass really needs the exact view type, it can downcast safely, since it was responsible for creating the view in the first place. In theory this could lead to faraway design issues like something depending _on_ a `ViewMenu` that is sensitive to the view type, but in practice this has never happened and is unlikely to ever happen with Framework usage where the view type is abstracted anyway.

Unfortunately this is a breaking change for legacy users of the Core Library, but should be easy to fix as it is just removing the generic.