Add registration of custom data, currently for button sprite maps.
While the implementation is somewhat of a kludge right now, it should provide the ability to add new data types relatively easily in the future without requiring a clumsy system of separate directories for every single type, and separate caches and retrieval methods for every type.

The categorization is now views + sprites + (custom) data and everything new will go in data.