Another minor tweak to Menu Slot fixed edges.
Previous value was actually too small even for vanilla UI (the right border with is 9 px, not 8 px) and a few more px are needed to prevent Clear Glasses from making a mess.

The sprite is 64px wide so this should be fine, as long as it doesn't need to render in a horizontal space smaller than 12px.