Add a Segmented Control (AKA Segmented Button).
This is more usable than dropdowns for small lists (<= 5 items) and also uses arbitrary children with selection index instead of relying on a generic definition. When used in the framework, the view model can write a trivial `SelectedValue` around the bound `SelectedIndex` property that accomplishes the same thing. Doing this allows non-text content, i.e. images, to be used.

Supports optional balancing of views so that all have the same width.

Reference docs are included here but user docs still need to be added to the Standard Views page.