Suppress mouse down/click events on newly-created overlays.
Without suppression, click events that cause an overlay to open can "leak" into that overlay on the initial frames. This is very easy to repro with the new color picker if the color wheel, which is a large target, intersects with the click location; it starts a drag even though the wheel was never technically clicked.

This appears to have always been an issue. Most overlays aren't affected because it's relatively uncommon for the overlay clickable area and original activation point to overlap (e.g. for a dropdown, it never will) and also, it mostly only affects drags, not conventional clicks, although suppressing is a good idea in both cases.

There may still be an issue for child menus but we can't really control that, it is up to the specific menu to handle its own suppression.