Add custom (named) state behaviors.
This allows a form of conditional attributes, by first associating a boolean condition with an arbitrary state name, and then using the state name to apply property changes (which can be transitioned, etc.).

Using this, we can associate a context property like "Selected" or "Enabled" to more view-centric concerns like "bold", "background tint", "opacity", etc. And can associate a single field with several of these view properties.

Fixes a bug in the `ViewBehavior` that could prevent the detach/reattach cycle from happening when the `ViewState` has been recreated but not the underlying View.