Support stretched dropdowns and multi-line drop-down options.
For stretching, it's impossible to resolve the layout ambiguity given a label/frame with stretched width and a button with stretched height, since the frame could _in theory_ (even if not in practice) change its height based on available width and the button actually does change its width based on accumulated height.

A workaround for now is to use a pixel height for the button, computed based on the font, which right now isn't allowed to be changed, so the calculation should be stable and accurate.