Allow button press handlers to short-circuit focus searches.
This is useful for keeping focus trapped in a particular area, e.g. when using directional buttons to navigate items in a list, move an element or slider, etc.

While `Slider` has its own ability to lock focus, other views created by the Framework don't have access to these internals and need to rely on events.

We could alternatively add a property for trapping focus, but that would increase the complexity of the API and is less generally applicable.