Allow non-dirty layout to propagate recursively to floating elements.
This fixes an obscure bug in which floating elements deeper in the hierarchy would not receive layout even when dirty. The exact scenario is:

1. Create a regular hierarchy, such as an inner frame within an outer frame.
2. Add a floating element to the inner view.
3. Make a change to something inside the floating element (discovery was changing the sprite of an `Image`).
4. Observe that `OnMeasure` doesn't run at the expected time.

Sometimes `OnMeasure` would run much later, I didn't investigate the exact trigger. The root cause was that non-dirty views were only looking at _their own_ direct floating elements, and since the layout intentionally doesn't propagate to children, the floats of children were missed.

The fix is to have a special pass just for floating elements, which are treated differently since they ignore normal flow layout and in this case we know the parents cannot be dirty (since we already checked).